Created HistoricEvent.Swift file and added in a list of events, a swift struct for getting a random event each time and a class for assigning a random event four times.
Added in the labels, buttons and countdown timer which doesn't yet function. The labels can fit four lines of text as of now which is all that's needed.

In the view controller I've created one main function for all the buttons and connected them all, they are differentiated using the sender.tag and I've assigned each button a unique tag. I also have two variables linking to the HistoricEvent.swift file to get four random events once the app loads.